requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers    
faiss-cpu                
numpy
requests
networkx                 
langdetect               
openai                   
unidecode                